test: unify scaffold
