--- packages/integration-tests/jest.config.ts
@@ -1,8 +1,0 @@
-import buildConfig from '../../jest.config'
-import type { Config } from '@jest/types'
-
-const config: Config.InitialOptions = {
-  ...buildConfig(__dirname),
-  testEnvironment: '@edge-runtime/jest-environment',
-}
-export default config
