--- packages/integration-tests/tests/crypto.test.ts
@@ -1,5 +1,0 @@
-import { crypto } from '@edge-runtime/ponyfill'
-
-test('crypto.randomUUID', () => {
-  expect(crypto.randomUUID()).toEqual(expect.stringMatching(/^[a-f0-9-]+$/))
-})
