+++ packages/vm/tests/integration/encoding.test.ts
@@ -1,4 +1,6 @@
-import { atob, btoa, TextDecoder } from '@edge-runtime/ponyfill'
+/**
+ * @jest-environment ../jest-environment/dist
+ */
 
 test('TextDecoder', () => {
   const input = new Uint8Array([
